fix: recommended subscriptions
