+++ src/ui/page/following/index.js
@@ -1,16 +1,19 @@
 import { connect } from 'react-redux';
 import { selectFollowedTags } from 'lbry-redux';
-import { selectSubscriptions } from 'redux/selectors/subscriptions';
+import { selectSubscriptions, selectSuggestedChannels } from 'redux/selectors/subscriptions';
+import { doFetchRecommendedSubscriptions } from 'redux/actions/subscriptions';
+
 import TagsEdit from './view';
 
 const select = state => ({
   followedTags: selectFollowedTags(state),
   subscribedChannels: selectSubscriptions(state),
+  suggestedSubscriptions: selectSuggestedChannels(state),
 });
-
-const perform = {};
 
 export default connect(
   select,
-  perform
+  {
+    doFetchRecommendedSubscriptions,
+  }
 )(TagsEdit);
