+++ src/ui/page/following/view.jsx
@@ -1,26 +1,61 @@
 // @flow
-import React from 'react';
+import * as PAGES from 'constants/pages';
+import React, { useEffect } from 'react';
 import Page from 'component/page';
 import TagsSelect from 'component/tagsSelect';
 import ClaimList from 'component/claimList';
+import Button from 'component/button';
 
 type Props = {
   subscribedChannels: Array<Subscription>,
+  location: { search: string },
+  history: { push: string => void },
+  doFetchRecommendedSubscriptions: () => void,
+  suggestedSubscriptions: Array<{ uri: string }>,
 };
 
-function FollowingEditPage(props: Props) {
-  const { subscribedChannels } = props;
+function FollowingPage(props: Props) {
+  const { subscribedChannels, location, history, doFetchRecommendedSubscriptions, suggestedSubscriptions } = props;
+  const hasSubscriptions = !!subscribedChannels.length;
   const channelUris = subscribedChannels.map(({ uri }) => uri);
+
+  const { search } = location;
+  const urlParams = new URLSearchParams(search);
+  const viewingSuggestedSubs = urlParams.get('view') || !hasSubscriptions;
+
+  function onClick() {
+    let url = `/$/${PAGES.FOLLOWING}`;
+    if (!viewingSuggestedSubs) {
+      url += '?view=discover';
+    }
+
+    history.push(url);
+  }
+
+  useEffect(() => {
+    doFetchRecommendedSubscriptions();
+  }, [doFetchRecommendedSubscriptions]);
+
   return (
     <Page>
       <div className="card">
         <TagsSelect showClose={false} title={__('Customize Your Tags')} />
       </div>
       <div className="card">
-        <ClaimList header={<h1>{__('Channels You Follow')}</h1>} uris={channelUris} />
+        <ClaimList
+          header={<h1>{viewingSuggestedSubs ? __('Discover New Channels') : __('Channels You Follow')}</h1>}
+          headerAltControls={
+            <Button
+              button="link"
+              label={viewingSuggestedSubs ? hasSubscriptions && __('View Your Channels') : __('Find New Channels')}
+              onClick={() => onClick()}
+            />
+          }
+          uris={viewingSuggestedSubs ? suggestedSubscriptions.map(sub => sub.uri) : channelUris}
+        />
       </div>
     </Page>
   );
 }
 
-export default FollowingEditPage;
+export default FollowingPage;
