+++ static/locales/en.json
@@ -480,5 +480,56 @@
   "Got it!": "Got it!",
   "Filter": "Filter",
   "Rendering document.": "Rendering document.",
-  "Sorry, looks like we can't load the document.": "Sorry, looks like we can't load the document."
+  "Sorry, looks like we can't load the document.": "Sorry, looks like we can't load the document.",
+  "Tag Search": "Tag Search",
+  "Disabled": "Disabled",
+  "Rewards are currently disabled for your account. Turn on diagnostic data sharing, in": "Rewards are currently disabled for your account. Turn on diagnostic data sharing, in",
+  ", in order to re-enable them.": ", in order to re-enable them.",
+  "Humans Only": "Humans Only",
+  "Rewards are for human beings only.": "Rewards are for human beings only.",
+  "You'll have to prove you're one of us before you can claim any rewards.": "You'll have to prove you're one of us before you can claim any rewards.",
+  "Rewawrds ": "Rewawrds ",
+  "Verification For Rewards": "Verification For Rewards",
+  "Fetching rewards": "Fetching rewards",
+  "This is optional.": "This is optional.",
+  "Verify Your Email": "Verify Your Email",
+  "Final Human Proof": "Final Human Proof",
+  "1) Proof via Credit": "1) Proof via Credit",
+  "If you have a valid credit or debit card, you can use it to instantly prove your humanity.": "If you have a valid credit or debit card, you can use it to instantly prove your humanity.",
+  "LBRY does not store your credit card information. There is no charge at all for this, now or in the future.": "LBRY does not store your credit card information. There is no charge at all for this, now or in the future.",
+  "Perform Card Verification": "Perform Card Verification",
+  "A $1 authorization may temporarily appear with your provider.": "A $1 authorization may temporarily appear with your provider.",
+  "Read more about why we do this.": "Read more about why we do this.",
+  "2) Proof via Phone": "2) Proof via Phone",
+  "You will receive an SMS text message confirming that your phone number is correct.": "You will receive an SMS text message confirming that your phone number is correct.",
+  "Submit Phone Number": "Submit Phone Number",
+  "Standard messaging rates apply. Having trouble?": "Standard messaging rates apply. Having trouble?",
+  "Read more.": "Read more.",
+  "3) Proof via Chat": "3) Proof via Chat",
+  "A moderator capable of approving you is typically available in the discord server. Check out the #rewards-approval channel for more information.": "A moderator capable of approving you is typically available in the discord server. Check out the #rewards-approval channel for more information.",
+  "This process will likely involve providing proof of a stable and established online or real-life identity.": "This process will likely involve providing proof of a stable and established online or real-life identity.",
+  "Join LBRY Chat": "Join LBRY Chat",
+  "Or, Skip It Entirely": "Or, Skip It Entirely",
+  "You can continue without this step, but you will not be eligible to earn rewards.": "You can continue without this step, but you will not be eligible to earn rewards.",
+  "Skip Rewards": "Skip Rewards",
+  "Waiting For Verification": "Waiting For Verification",
+  "An email was sent to": "An email was sent to",
+  "Follow the link and you will be good to go. This will update automatically.": "Follow the link and you will be good to go. This will update automatically.",
+  "Resend verification email": "Resend verification email",
+  "if you encounter any trouble verifying.": "if you encounter any trouble verifying.",
+  "Blockchain Sync": "Blockchain Sync",
+  "Catching up with the blockchain": "Catching up with the blockchain",
+  "No Rewards Available": "No Rewards Available",
+  "There are no rewards available at this time, please check back later.": "There are no rewards available at this time, please check back later.",
+  "Confirm Identity": "Confirm Identity",
+  "Not following any channels": "Not following any channels",
+  "Subscriptions 101": "Subscriptions 101",
+  "You just subscribed to your first channel. Awesome!": "You just subscribed to your first channel. Awesome!",
+  "A few quick things to know:": "A few quick things to know:",
+  "1) This app will automatically download new free content from channels you are subscribed to. You may configure this in Settings or on the Subscriptions page.": "1) This app will automatically download new free content from channels you are subscribed to. You may configure this in Settings or on the Subscriptions page.",
+  "2) If we have your email address, we will send you notifications related to new content. You may configure these emails from the Help page.": "2) If we have your email address, we will send you notifications related to new content. You may configure these emails from the Help page.",
+  "Got it": "Got it",
+  "View Your Feed": "View Your Feed",
+  "View Your Channels": "View Your Channels",
+  "Unfollow": "Unfollow"
 }
